requirements.txt
telethon==1.36.0
groq
python-dotenv==1.0.1
aiohttp==3.10.10
pillow==10.4.0
